refactor(middleware): extract the delegation gate + verification plumbing
Move wrapWithDelegation into with-kya-os.delegation-gate.ts (the request flow) and the DID resolver / embedded-proof signature verifier / chain-validation adapter / needs_authorization challenge builder into with-kya-os.delegation-verify.ts, wired via createDelegationGate(deps, { attachOutcomeProof, resolveExistingGrant, bindGrantOnSuccess }). Prune the now-dead delegation imports from the orchestrator. No behaviour change; 1607 tests green. Phase 2c.
