+++ src/middleware/with-kya-os.ts
@@ -15,17 +15,13 @@
  * low-level `Server` API or custom request handler patterns.
  */
 
-import {
-  type CryptoProvider,
-  type FetchProvider,
-  type NonceCacheProvider,
-} from "../providers/base.js";
+import { type CryptoProvider } from "../providers/base.js";
 import { RuntimeFetchProvider, NoopFetchProvider } from "../providers/runtime-fetch.js";
-import { AuditLogProvider, NoopAuditLogProvider } from "../providers/audit-log.js";
-import { GrantStore, MemoryGrantStore, type Grant } from "../providers/grant-store.js";
+import { NoopAuditLogProvider } from "../providers/audit-log.js";
+import { GrantStore, MemoryGrantStore } from "../providers/grant-store.js";
 import {
   SessionManager,
-  type SessionConfig,
+  validateHandshakeFormat,
   type HandshakeResult,
 } from "../session/manager.js";
 import {
@@ -36,45 +32,13 @@ import {
   type ToolRequest,
   type ToolResponse,
 } from "../proof/generator.js";
-import { validateHandshakeFormat } from "../session/manager.js";
-import {
-  DelegationCredentialVerifier,
-  type DIDResolver,
-  type SignatureVerificationFunction,
-  type StatusListResolver,
-} from "../delegation/vc-verifier.js";
-import { createDidKeyResolver } from "../delegation/did-key-resolver.js";
-import {
-  assertHolderBinding,
-  isHolderBindingApplicable,
-  isKyaOsControlArg,
-  toHolderBindingRequest,
-} from "../delegation/holder-binding.js";
 import { ProofVerifier } from "../proof/verifier.js";
 import { SystemClockProvider } from "../providers/system-clock.js";
 import { MemoryNonceCacheProvider } from "../providers/memory.js";
-import { scopeSatisfies } from "../delegation/scope-matcher.js";
-import { createDidWebResolver } from "../delegation/did-web-resolver.js";
-import {
-  buildDidResolverRegistry,
-  type DIDResolverRegistry,
-} from "../delegation/did-resolver-registry.js";
-import { getDidMethod } from "../utils/did-helpers.js";
-import {
-  validateDelegationChain as validateDelegationChainCore,
-  getDelegationScopes,
-  type RevocationChecker,
-} from "../delegation/chain-enforcement.js";
-import {
-  createNeedsAuthorizationError,
-  type DelegationCredential,
-  type DetachedProof,
-  type NeedsAuthorizationError,
-} from "../types/protocol.js";
+import type { DetachedProof } from "../types/protocol.js";
 import { logger } from "../logging/index.js";
 import { KYA_OS_ERROR_CODES } from "../errors.js";
-import { canonicalizeJSON, parseVCJWT } from "../delegation/utils.js";
-import { base64urlDecodeToBytes, base64urlEncodeFromBytes, bytesToBase64 } from "../utils/base64.js";
+import { base64urlEncodeFromBytes } from "../utils/base64.js";
 import { sanitizeForMessage } from "./with-kya-os.helpers.js";
 import {
   KYA_OS_ACTIONS,
@@ -84,11 +48,11 @@ import {
   type KyaOsToolDefinition,
   type KyaOsCallContext,
   type KyaOsMiddleware,
-  type PolicyGateOptions,
 } from "./with-kya-os.types.js";
 import type { MiddlewareDeps } from "./with-kya-os.deps.js";
 import { createGrantResolution } from "./with-kya-os.grants.js";
 import { createPolicyGate } from "./with-kya-os.policy-gate.js";
+import { createDelegationGate } from "./with-kya-os.delegation-gate.js";
 
 // The public type surface lives in `./with-kya-os.types.ts`; re-export it here so
 // `import { KyaOsConfig, ... } from './with-kya-os'` continues to resolve unchanged.
@@ -588,420 +552,11 @@ export function createKyaOsMiddleware(
     return response;
   }
 
-  function wrapWithDelegation(
-    toolName: string,
-    config: {
-      scopeId: string;
-      consentUrl: string;
-      formatChallenge?: (
-        challenge: NeedsAuthorizationError,
-      ) => Array<{ type: "text"; text: string }>;
-    },
-    handler: KyaOsToolHandler,
-  ): KyaOsToolHandler {
-    const didKeyResolver = createDidKeyResolver();
-    const fetchProvider =
-      delegationConfig?.fetchProvider ??
-      (typeof globalThis.fetch === "function"
-        ? new RuntimeFetchProvider()
-        : undefined);
-    const didWebResolver = fetchProvider
-      ? createDidWebResolver(fetchProvider)
-      : undefined;
-    const configuredDidResolvers = buildDidResolverRegistry(
-      delegationConfig?.didResolvers,
-      fetchProvider,
-    );
-    const didResolver: DIDResolver = {
-      async resolve(did: string) {
-        const customResolver = delegationConfig?.didResolver;
-        if (customResolver) {
-          const resolved = await customResolver.resolve(did);
-          if (resolved) {
-            return resolved;
-          }
-        }
-
-        const method = getDidMethod(did);
-        const configuredResolver = method ? configuredDidResolvers[method] : undefined;
-        if (configuredResolver) {
-          try {
-            const resolved = await configuredResolver.resolve(did);
-            if (resolved) {
-              return resolved;
-            }
-          } catch {
-            return null;
-          }
-        }
-
-        if (did.startsWith("did:key:")) {
-          return didKeyResolver.resolve(did);
-        }
-
-        if (did.startsWith("did:web:")) {
-          return didWebResolver?.resolve(did) ?? null;
-        }
-
-        return null;
-      },
-    };
-
-    const signatureVerifier: SignatureVerificationFunction = async (
-      vc: DelegationCredential,
-      publicKeyJwk: unknown,
-    ): Promise<{ valid: boolean; reason?: string }> => {
-      const proof = vc.proof;
-      if (!proof) {
-        return { valid: false, reason: "Missing proof" };
-      }
-
-      const proofValue = proof["proofValue"] as string | undefined;
-      if (!proofValue) {
-        return { valid: false, reason: "Missing proofValue in proof" };
-      }
-
-      // Reconstruct the unsigned VC (without proof) for signature verification
-      const vcRecord = vc as Record<string, unknown>;
-      const vcWithoutProof: Record<string, unknown> = {};
-      for (const [k, v] of Object.entries(vcRecord)) {
-        if (k !== "proof") vcWithoutProof[k] = v;
-      }
-      const canonical = canonicalizeJSON(vcWithoutProof);
-      const data = new TextEncoder().encode(canonical);
-
-      // Decode signature from base64url proof value
-      const sigBytes = base64urlDecodeToBytes(proofValue);
-
-      // Get public key from JWK (x is base64url-encoded raw key bytes)
-      const jwk = publicKeyJwk as { x?: string };
-      if (!jwk.x) {
-        return { valid: false, reason: "No x field in publicKeyJwk" };
-      }
-
-      // Convert base64url key to standard base64 for the crypto provider
-      const pubKeyBytes = base64urlDecodeToBytes(jwk.x);
-      const pubKeyBase64 = bytesToBase64(pubKeyBytes);
-
-      const valid = await cryptoProvider.verify(data, sigBytes, pubKeyBase64);
-      return {
-        valid,
-        reason: valid ? undefined : "Signature verification failed",
-      };
-    };
-
-    const verifier = new DelegationCredentialVerifier({
-      didResolver,
-      signatureVerifier,
-      statusListResolver: delegationConfig?.statusListResolver,
-    });
-
-    const buildDelegationErrorResponse = (
-      error: string,
-      reason: string,
-    ): Awaited<ReturnType<KyaOsToolHandler>> => ({
-      content: [
-        {
-          type: "text" as const,
-          text: JSON.stringify({ error, reason: sanitizeForMessage(reason) }),
-        },
-      ],
-      isError: true,
-    });
-
-    // Thin adapter over the framework-agnostic core (E3.1 · chain-enforcement.ts).
-    // The walk/attenuation/audience/§11.6/revocation rules live in one place;
-    // this binds the host's injected verifier + server DID + resolver + the
-    // optional graph-backed RevocationChecker.
-    const validateDelegationChain = (
-      leafCredential: DelegationCredential,
-      options?: { skipSignature?: boolean },
-    ): Promise<{ valid: boolean; reason?: string }> =>
-      validateDelegationChainCore(
-        leafCredential,
-        {
-          serverDid: identity.did,
-          verifier,
-          resolveDelegationChain: delegationConfig?.resolveDelegationChain,
-          statusListConfigured: !!delegationConfig?.statusListResolver,
-          revocationChecker: delegationConfig?.revocationChecker,
-        },
-        options,
-      );
-
-    return async (
-      args: Record<string, unknown>,
-      sessionId?: string,
-    ) => {
-      const delegationArg = args["_kyaos_delegation"];
-
-      if (delegationArg === undefined || delegationArg === null) {
-        // No delegation pasted — a durable grant may already authorize this call
-        // (the no-paste retry), even on a fresh instance with empty memory.
-        // Holder-of-key first (agent-anchored, proof-gated), then the session
-        // bearer capability. On a hit, skip the challenge and run the handler
-        // with exactly the call shape a verified delegation would have produced.
-        const existingGrant = await resolveExistingGrant(
-          toolName,
-          args,
-          sessionId,
-          config.scopeId,
-        );
-        if (existingGrant) {
-          const grantArgs: Record<string, unknown> = {};
-          for (const [k, v] of Object.entries(args)) {
-            if (!isKyaOsControlArg(k)) grantArgs[k] = v;
-          }
-          logger.debug(
-            `[kya-os] Grant resolved for "${toolName}" (scope "${config.scopeId}") — no re-paste required`,
-          );
-          return handler(grantArgs, sessionId, { scopeId: config.scopeId });
-        }
-
-        // No delegation provided — return needs_authorization response
-        const tokenBytes = await cryptoProvider.randomBytes(16);
-        const hex = Array.from(tokenBytes)
-          .map((b) => b.toString(16).padStart(2, "0"))
-          .join("");
-        const resumeToken = [
-          hex.slice(0, 8),
-          hex.slice(8, 12),
-          hex.slice(12, 16),
-          hex.slice(16, 20),
-          hex.slice(20),
-        ].join("-");
-        const expiresAt = Math.floor(Date.now() / 1000) + 300;
-
-        const authError = createNeedsAuthorizationError({
-          message: `Tool "${toolName}" requires delegation with scope: ${config.scopeId}`,
-          authorizationUrl: config.consentUrl,
-          resumeToken,
-          expiresAt,
-          scopes: [config.scopeId],
-        });
-
-        // Sign the challenge (outcome=needs_authorization). The proof binds a
-        // responseHash over the EMITTED challenge content — including the
-        // authorizationUrl. A verifier that recomputes the response hash over the
-        // content it received (ProofVerifier.verifyProof(proof, jwk, { request,
-        // response })) thereby detects a tampered/MITM-swapped consent URL; the
-        // signature alone proves authenticity, not content-match. config.format-
-        // Challenge lets a server render the challenge (e.g. a markdown link for
-        // LLM clients) BEFORE signing, so the proof binds exactly what the client
-        // receives. A throwing hook falls back to the default challenge (never
-        // -32603). Best-effort: attachOutcomeProof no-ops if no session resolves.
-        const defaultChallengeContent = [
-          { type: "text" as const, text: JSON.stringify(authError) },
-        ];
-        let challengeContent = defaultChallengeContent;
-        if (config.formatChallenge) {
-          try {
-            challengeContent = config.formatChallenge(authError);
-          } catch (error) {
-            logger.error("[kya-os] formatChallenge threw; using the default challenge", {
-              tool: toolName,
-              error: error instanceof Error ? error.message : String(error),
-            });
-            challengeContent = defaultChallengeContent;
-          }
-        }
-        return attachOutcomeProof(
-          { content: challengeContent },
-          toolName,
-          args,
-          sessionId,
-          authError.message,
-          "needs_authorization",
-          undefined,
-          challengeContent,
-        );
-      }
-
-      // Accept delegation as either a JSON object (embedded proof) or a
-      // VC-JWT string (compact JWT). The Cloudflare consent service issues
-      // VC-JWTs; examples use embedded proofs. Support both transparently.
-      let vc: DelegationCredential;
-      let isVCJWT = false;
-      if (typeof delegationArg === "string") {
-        const parsed = parseVCJWT(delegationArg);
-        if (!parsed || !parsed.payload.vc) {
-          return attachOutcomeProof(
-            buildDelegationErrorResponse(
-              KYA_OS_ERROR_CODES.delegation_invalid,
-              "Invalid VC-JWT format",
-            ),
-            toolName,
-            args,
-            sessionId,
-            "Invalid VC-JWT format",
-          );
-        }
-        vc = parsed.payload.vc as DelegationCredential;
-        // VC-JWTs don't have an embedded proof — the JWT signature is the
-        // proof. Add a marker so basic validation (which checks for proof
-        // presence) passes. The actual signature is in the JWT envelope.
-        if (!vc.proof) {
-          vc = { ...vc, proof: { type: 'JwtProof2020', jwt: delegationArg } };
-        }
-        isVCJWT = true;
-      } else {
-        vc = delegationArg as DelegationCredential;
-      }
-
-      // For VC-JWTs the embedded-signature check is skipped (the JWT envelope
-      // signature is the proof); schema/expiry/status/scope checks still apply.
-      // validateDelegationChain performs the shape check and returns
-      // { valid, reason } for malformed input, never throwing on normal or
-      // structurally-malformed credentials. This try/catch is therefore a PURE
-      // BACKSTOP — it fires only on a truly unexpected throw (e.g. a hostile
-      // getter/Proxy accessor or a provider fault): it logs the detail
-      // server-side and returns a GENERIC reason, so no internal/implementation
-      // detail (stack, raw error text) leaks to the client or the signed proof.
-      const verificationResult = await (async (): Promise<{
-        valid: boolean;
-        reason?: string;
-      }> => {
-        try {
-          return await validateDelegationChain(vc, { skipSignature: isVCJWT });
-        } catch (error) {
-          logger.error("[kya-os] Unexpected error verifying delegation", {
-            tool: toolName,
-            error: error instanceof Error ? error.message : String(error),
-            stack: error instanceof Error ? error.stack : undefined,
-          });
-          return { valid: false, reason: "Delegation credential could not be verified" };
-        }
-      })();
-
-      if (!verificationResult.valid) {
-        const reason = verificationResult.reason ?? "Unknown delegation validation error";
-        logger.warn(
-          `[kya-os] Delegation verification failed for "${toolName}": ${sanitizeForMessage(reason)}`,
-        );
-        return attachOutcomeProof(
-          buildDelegationErrorResponse(KYA_OS_ERROR_CODES.delegation_invalid, reason),
-          toolName,
-          args,
-          sessionId,
-          reason,
-        );
-      }
-
-      // Holder binding (spec §11.8): the delegation is valid, but a valid
-      // *credential* is a bearer token until we also prove the caller holds the
-      // delegation SUBJECT's key. Opt-in via `delegation.holderBinding`. did:key
-      // subjects are bound here; did:web is deferred to cnf binding (phase 2) and
-      // logged, never rejected. Runs after identity is established, before scope.
-      if (holderBindingMode !== "off" && holderBindingVerifier) {
-        const subjectDid = vc.credentialSubject?.id;
-        if (subjectDid && isHolderBindingApplicable(subjectDid)) {
-          const proofArg = args["_kyaos_proof"];
-          if (proofArg === undefined) {
-            const reason =
-              "Holder-of-key proof (_kyaos_proof) is required for this delegation subject";
-            logger.warn(
-              `[kya-os] Holder binding: "${toolName}" called without _kyaos_proof`,
-            );
-            if (holderBindingMode === "enforce") {
-              return attachOutcomeProof(
-                buildDelegationErrorResponse(
-                  KYA_OS_ERROR_CODES.holder_binding_failed,
-                  reason,
-                ),
-                toolName,
-                args,
-                sessionId,
-                reason,
-              );
-            }
-          } else {
-            let parsedProof: unknown = proofArg;
-            if (typeof proofArg === "string") {
-              try {
-                parsedProof = JSON.parse(proofArg);
-              } catch {
-                parsedProof = {};
-              }
-            }
-            const binding = await assertHolderBinding({
-              proof: parsedProof as DetachedProof,
-              subjectDid,
-              request: toHolderBindingRequest(toolName, args),
-              expectedAudience: identity.did,
-              proofVerifier: holderBindingVerifier,
-            });
-            if (binding.status !== "bound") {
-              const reason =
-                binding.reason ??
-                "Holder-of-key proof did not bind the delegation subject";
-              logger.warn(
-                `[kya-os] Holder binding ${binding.status} for "${toolName}": ${sanitizeForMessage(reason)}`,
-              );
-              if (holderBindingMode === "enforce") {
-                return attachOutcomeProof(
-                  buildDelegationErrorResponse(
-                    KYA_OS_ERROR_CODES.holder_binding_failed,
-                    reason,
-                  ),
-                  toolName,
-                  args,
-                  sessionId,
-                  reason,
-                );
-              }
-            }
-          }
-        } else if (subjectDid) {
-          // Non-did:key subject — phase 1 cannot pin its key; defer to cnf
-          // binding (phase 2) rather than reject legitimate traffic.
-          logger.warn(
-            `[kya-os] Holder binding: subject "${subjectDid}" is not did:key; deferring to cnf binding (phase 2)`,
-          );
-        }
-      }
-
-      // Safe to call directly: the structural guard + validateDelegationChain
-      // above guarantee a well-formed credential here, and scopeSatisfies is
-      // bounded (ReDoS-guarded) and returns rather than throws.
-      const scopeResult = scopeSatisfies(config.scopeId, vc);
-      if (scopeResult.usedNonExactMatcher) {
-        logger.warn(
-          `[kya-os] Scope "${config.scopeId}" for "${toolName}" granted via a non-exact ` +
-            `(prefix/regex) matcher. Verify this is intended — non-exact matchers widen authority.`,
-        );
-      }
-      if (!scopeResult.satisfied) {
-        const reason = `Required scope "${config.scopeId}" not in delegation scopes`;
-        logger.warn(
-          `[kya-os] Delegation missing required scope "${config.scopeId}" for "${toolName}"`,
-        );
-        return attachOutcomeProof(
-          buildDelegationErrorResponse(KYA_OS_ERROR_CODES.insufficient_scope, reason),
-          toolName,
-          args,
-          sessionId,
-          reason,
-        );
-      }
-
-      // Strip the reserved _kyaos* control namespace before passing to the
-      // handler — same predicate the bound request hash uses, so the handler
-      // receives exactly the call the subject signed (no smuggled control arg).
-      const cleanArgs: Record<string, unknown> = {};
-      for (const [k, v] of Object.entries(args)) {
-        if (!isKyaOsControlArg(k)) cleanArgs[k] = v;
-      }
-
-      // Mint a durable grant from this verified delegation so the next call —
-      // on any instance — resolves via resolveExistingGrant with no re-paste.
-      await bindGrantOnSuccess(vc, delegationArg, isVCJWT, sessionId, config.scopeId);
-
-      logger.debug(
-        `[kya-os] Delegation verified for "${toolName}", scope "${config.scopeId}"`,
-      );
-      return handler(cleanArgs, sessionId, { scopeId: config.scopeId });
-    };
-  }
+  const { wrapWithDelegation } = createDelegationGate(deps, {
+    attachOutcomeProof,
+    resolveExistingGrant,
+    bindGrantOnSuccess,
+  });
 
   const { withPolicyGate } = createPolicyGate(deps, { attachOutcomeProof });
